added organization logic, code refactoring

+++ TaskManagerApi/Controllers/TaskItemController.cs
@@ -7,7 +7,9 @@ using TaskManagerApi.Data;
 using TaskManagerApi.Enitities;
 using TaskManagerApi.Models;
 using TaskManagerApi.Models.TaskItem;
+using TaskManagerApi.Services.Implementations;
 using TaskManagerApi.Services.Interfaces;
+using static TaskManagerApi.Models.Constants;
 
 namespace TaskManagerApi.Controllers
 {
@@ -17,14 +19,22 @@ namespace TaskManagerApi.Controllers
     public class TaskItemController(ITaskItemService taskItemService) : ControllerBase
     {
         [HttpGet("all")]
-        public async Task<ActionResult<List<TaskItemOutputDto>>> GetTasks()
+        public async Task<ActionResult<List<TaskItemDto>>> GetTasksInOrganization()
         {
-            return Ok(await taskItemService.GetTasksAsync());
+            if(!this.Request.Headers.TryGetValue("organizationId", out var organizationId) 
+                || !await ValidateAccountOrganizationConnection(organizationId!)) 
+                return BadRequest("Invalid request");
+
+            return Ok(await taskItemService.GetTasksByOrganizationAsync());
         }
 
         [HttpGet("details/{Id}")]
-        public async Task<ActionResult<TaskItemOutputDto>> GetTaskById(Guid Id)
+        public async Task<ActionResult<TaskItemDto>> GetTaskById(Guid Id)
         {
+            if(!this.Request.Headers.TryGetValue("organizationId", out var organizationId) 
+                || !await ValidateAccountOrganizationConnection(organizationId!)) 
+                return BadRequest("Invalid request");
+
             var task = await taskItemService.GetTaskByIdAsync(Id);
 
             if (task is null)
@@ -34,8 +44,12 @@ namespace TaskManagerApi.Controllers
         }
 
         [HttpPost("create")]
-        public async Task<ActionResult<TaskItemOutputDto>> CreateTask(TaskItemInputDto newTask)
+        public async Task<ActionResult<TaskItemDto>> CreateTask(TaskItemDto newTask)
         {
+            if(!this.Request.Headers.TryGetValue("organizationId", out var organizationId) 
+                || !await ValidateAccountOrganizationConnection(organizationId!)) 
+                return BadRequest("Invalid request");
+
             var createdTask = await taskItemService.CreateTaskAsync(newTask);
 
             if (createdTask is null)
@@ -44,9 +58,13 @@ namespace TaskManagerApi.Controllers
             return Ok(createdTask);
         }
 
-        [HttpPut("edit/{Id}")]
-        public async Task<ActionResult<TaskItemOutputDto>> EditTaskById(Guid Id, [FromBody] TaskItemInputDto editTask)
+        [HttpPost("edit/{Id}")]
+        public async Task<ActionResult<TaskItemDto>> EditTaskById(Guid Id, TaskItemDto editTask)
         {
+            if(!this.Request.Headers.TryGetValue("organizationId", out var organizationId) 
+                || !await ValidateAccountOrganizationConnection(organizationId!)) 
+                return BadRequest("Invalid request");
+
             var taskToEdit = await taskItemService.EditTaskByIdAsync(Id, editTask);
 
             if (taskToEdit is null)
@@ -56,8 +74,12 @@ namespace TaskManagerApi.Controllers
         }
 
         [HttpPut("edit/{taskId}/assignee/{assigneeId}")]
-        public async Task<ActionResult<TaskItemOutputDto>> ChangeTaskAssigneeById(Guid taskId, Guid assigneeId)
+        public async Task<ActionResult<TaskItemDto>> ChangeTaskAssigneeById(Guid taskId, Guid assigneeId)
         {
+            if(!this.Request.Headers.TryGetValue("organizationId", out var organizationId) 
+                || !await ValidateAccountOrganizationConnection(organizationId!)) 
+                return BadRequest("Invalid request");
+
             var taskToEdit = await taskItemService.ChangeAssigneeAsync(taskId, assigneeId);
 
             if (taskToEdit is null)
@@ -67,8 +89,12 @@ namespace TaskManagerApi.Controllers
         }
 
         [HttpPut("edit/{taskId}/project/{projectId}")]
-        public async Task<ActionResult<TaskItemOutputDto>> ChangeTaskProject(Guid Id, Guid projectId)
+        public async Task<ActionResult<TaskItemDto>> ChangeTaskProject(Guid Id, Guid projectId)
         {
+            if(!this.Request.Headers.TryGetValue("organizationId", out var organizationId) 
+                || !await ValidateAccountOrganizationConnection(organizationId!)) 
+                return BadRequest("Invalid request");
+
             var taskToEdit = await taskItemService.ChangeAssigneeAsync(Id, projectId);
 
             if (taskToEdit is null)
@@ -78,8 +104,12 @@ namespace TaskManagerApi.Controllers
         }
 
         [HttpPut("edit/{taskId}/task/{statusId}")]
-        public async Task<ActionResult<TaskItemOutputDto>> ChangeTaskStatus(Guid Id, int statusId)
+        public async Task<ActionResult<TaskItemDto>> ChangeTaskStatus(Guid Id, int statusId)
         {
+            if(!this.Request.Headers.TryGetValue("organizationId", out var organizationId) 
+                || !await ValidateAccountOrganizationConnection(organizationId!)) 
+                return BadRequest("Invalid request");
+
             var taskToEdit = await taskItemService.ChangeTaskStatusAsync(Id, statusId);
 
             if (taskToEdit is null)
@@ -89,14 +119,25 @@ namespace TaskManagerApi.Controllers
         }
 
         [HttpDelete("delete/{Id}")]
-        public async Task<ActionResult<TaskItemOutputDto>> DeleteTaskById(Guid Id)
+        public async Task<ActionResult<TaskItemDto>> DeleteTaskById(Guid Id)
         {
+            if(!this.Request.Headers.TryGetValue("organizationId", out var organizationId) 
+                || !await ValidateAccountOrganizationConnection(organizationId!)) 
+                return BadRequest("Invalid request");
+
             var taskToDelete = await taskItemService.DeleteTaskAsync(Id);
 
             if (taskToDelete is null)
                 return NotFound();
 
             return Ok(taskToDelete);
+        }
+
+        private async Task<bool> ValidateAccountOrganizationConnection(string organizationId)
+        {
+            return (!Guid.TryParse(organizationId, out var organizationIdGuid)
+                || await GeneralService.VerifyAccountRelatesToOrganization(Guid.Parse(User.FindFirst(IdentityCustomOpenId.DetailsFromToken.ACCOUNT_ID)!.Value)
+                    , organizationIdGuid) is null);
         }
     }
 }
